refactor: 471 - move types & fix imports issues
