--- src/app/types/course-status.types.ts
@@ -1,5 +1,0 @@
-export enum CourseStatus {
-  PLANNED = 'planned',
-  AVAILABLE = 'available',
-  UPCOMING = 'upcoming',
-}
