--- src/app/types/course.types.ts
@@ -1,17 +1,0 @@
-export type Course = {
-  id: string;
-  title: string;
-  altTitle?: string;
-  iconSrc: string;
-  secondaryIcon: string;
-  iconSmall: string;
-  startDate: string;
-  language: ('en' | 'ru')[];
-  mode: 'online' | 'offline';
-  detailsUrl: string;
-  enroll: string;
-  backgroundStyle: {
-    backgroundColor: string;
-    accentColor: string;
-  };
-};
