--- src/app/types/index.ts
@@ -1,2 +1,0 @@
-export { type Course } from './course.types';
-export { CourseStatus } from './course-status.types';
