+++ src/entities/course/index.ts
@@ -1,1 +1,3 @@
+export type { Course, CourseCardProps } from './types';
 export { CourseCard } from './ui/course-card/course-card';
+export { CourseStatus } from './types';
