+++ src/entities/course/types.ts
@@ -1,1 +1,28 @@
-export type Day = number;
+export type Course = {
+  id: string;
+  title: string;
+  altTitle?: string;
+  iconSrc: string;
+  secondaryIcon: string;
+  iconSmall: string;
+  startDate: string;
+  language: ('en' | 'ru')[];
+  mode: 'online' | 'offline';
+  detailsUrl: string;
+  enroll: string;
+  backgroundStyle: {
+    backgroundColor: string;
+    accentColor: string;
+  };
+};
+
+export type CourseCardProps = Pick<
+  Course,
+  'title' | 'iconSrc' | 'startDate' | 'detailsUrl' | 'mode' | 'language' | 'backgroundStyle'
+>;
+
+export enum CourseStatus {
+  PLANNED = 'planned',
+  AVAILABLE = 'available',
+  UPCOMING = 'upcoming',
+}
