+++ src/entities/course/ui/course-card/course-card.tsx
@@ -1,15 +1,10 @@
-import { type Course } from '@/app/types';
+import type { CourseCardProps } from '../../types';
 import { ArrowIcon } from '@/shared/icons';
 import { DateLang } from '@/shared/ui/date-lang';
 import { Image } from '@/shared/ui/image';
 import { LinkCustom } from '@/shared/ui/link-custom';
 
 import './course-card.scss';
-
-export type CourseCardProps = Pick<
-  Course,
-  'title' | 'iconSrc' | 'startDate' | 'detailsUrl' | 'mode' | 'language' | 'backgroundStyle'
->;
 
 export const CourseCard = ({
   title,
