+++ src/shared/helpers/getActualData.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { getActualData } from './getActualData';
-import { Course } from '@/app/types';
+import type { Course } from '@/entities/course';
 import { EventCardProps } from '@/entities/events';
 import { dayJS } from '@/shared/helpers/dayJS';
 import { isCourse } from '@/shared/helpers/is-course';
