+++ src/shared/hooks/use-course-by-title/utils/is-course.ts
@@ -1,4 +1,4 @@
-import { Course } from '@/app/types';
+import type { Course } from '@/entities/course';
 
 const isCourse = (obj: object) => {
   return 'title' in obj && (obj as Course).title != null;
