+++ src/widgets/course-main/course-main.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from '@testing-library/react';
 import { CourseMain } from './course-main';
-import { CourseStatus } from '@/app/types';
+import { CourseStatus } from '@/entities/course';
 import { MOCKED_IMAGE_PATH } from '@/shared/__tests__/constants';
 import { renderWithRouter } from '@/shared/__tests__/utils';
 import { dayJS } from '@/shared/helpers/dayJS';
